1. optimze interpreter 2. optimze getThisBinding operation 3. optimze esprima::scanStringLiteral 4. optimze ObjectStructurePropertyDescriptor

+++ src/interpreter/ByteCodeGenerator.cpp
@@ -104,7 +104,7 @@ void ByteCodeGenerator::generateByteCode(Context* c, CodeBlock* codeBlock, Node*
 
 #ifndef NDEBUG
     if (getenv("DUMP_BYTECODE") && strlen(getenv("DUMP_BYTECODE"))) {
-        printf("dumpBytecode %s (%d:%d)>>>>>>>>>>>>>>>>>>>>>>\n", codeBlock->m_functionName.string()->toUTF8StringData().data(), codeBlock->sourceElementStart().line, codeBlock->sourceElementStart().column);
+        printf("dumpBytecode %s (%d:%d)>>>>>>>>>>>>>>>>>>>>>>\n", codeBlock->m_functionName.string()->toUTF8StringData().data(), (int)codeBlock->sourceElementStart().line, (int)codeBlock->sourceElementStart().column);
         size_t idx = 0;
         size_t bytecodeCounter = 0;
         char* code = block->m_code.data();
